Update component list to be shuffled
To enable the reproducible generation of a board using a random seed
this method to shuffle the list of components was implemented: Instead
of shuffling the whole list of components which could result in
inadequate or even unsolvable orders like [R6, R4, R5, B6, B5, ...],
each color is repeated in the same interval and only the order in which
the colors are picked and the order in which the component sizes are
picked are random. This seems to be adequate, since the algorithm for
filling a board works in columns from left to right and needs each color
in each column.